♻️ Refactor: mockData 수정, user.json 추가, index.d.ts로 이름 변환
