+++ src/components/content/Content.tsx
@@ -2,7 +2,7 @@ import { css } from '@emotion/react';
 
 import LeftBubble from './LeftBubble';
 import RightBubble from './RightBubble';
-import { Message } from '../../sns';
+import { Message } from '../..';
 
 interface ContentProps {
   isMuted: boolean;
