+++ src/mock/mockData.json
@@ -1,62 +1,52 @@
 [
-    {
-      "id": 1,
-      "status": 0,
-      "content": "땅을 보고 계속 올랐지",
-      "time": "오후 12:00"
-    },
-    {
-      "id": 2,
-      "status": 1,
-      "content": "정상까지",
-      "time": "오후 12:00"
-    },
-    {
-      "id": 3,
-      "status": 0,
-      "content": "많은 시련은 보란 듯이",
-      "time": "오후 12:00"
-    },
-    {
-      "id": 4,
-      "status": 1,
-      "content": "I always Win",
-      "time": "오후 12:00"
-    },
-    {
-      "id": 5,
-      "status": 0,
-      "content": "강한 맘이 중요하지",
-      "time": "오후 12:00"
-    },
-    {
-      "id": 6,
-      "status": 1,
-      "content": "미래는 도망가지 않아, \n내가 놓기 전까지",
-      "time": "오후 12:00"
-    },
-    {
-      "id": 7,
-      "status": 0,
-      "content": "마치 된 것 같아 손오공",
-      "time": "오후 12:00"
-    },
-    {
-      "id": 8,
-      "status": 1,
-      "content": "우리는 쉬지 않아 매일",
-      "time": "오후 12:00"
-    },
-    {
-      "id": 9,
-      "status": 0,
-      "content": "끝까지 가보자",
-      "time": "오후 12:00"
-    },
-    {
-      "id": 10,
-      "status": 1,
-      "content": "마치 된 것 같아 손오공",
-      "time": "오후 12:00"
-    }
+  {
+    "userId": 1,
+    "content": "땅을 보고 계속 올랐지",
+    "time": "오후 12:00"
+  },
+  {
+    "userId": 0,
+    "content": "정상까지",
+    "time": "오후 12:00"
+  },
+  {
+    "userId": 1,
+    "content": "많은 시련은 보란 듯이",
+    "time": "오후 12:00"
+  },
+  {
+    "userId": 0,
+    "content": "I always Win",
+    "time": "오후 12:00"
+  },
+  {
+    "userId": 1,
+    "content": "강한 맘이 중요하지",
+    "time": "오후 12:00"
+  },
+  {
+    "userId": 0,
+    "content": "미래는 도망가지 않아, \n내가 놓기 전까지",
+    "time": "오후 12:00"
+  },
+  {
+    "userId": 1,
+    "content": "마치 된 것 같아 손오공",
+    "time": "오후 12:00"
+  },
+  {
+    "userId": 0,
+    "content": "우리는 쉬지 않아 매일",
+    "time": "오후 12:00"
+  },
+  {
+    "userId": 1,
+    "content": "끝까지 가보자",
+    "time": "오후 12:00"
+  },
+  {
+    "userId": 0,
+    "content": "마치 된 것 같아 손오공",
+    "time": "오후 12:00"
+  }
 ]
